feat: add OrderValidator
This commit adds an OrderValidator contract which uses the OrderQuoter
to get generalized validation on order objects. The OrderQuoter
basically does a dry-run of an order execution, returning any errors
that came up during the processing of the order. This should act as
a catch-all to ensure orders are valid - it's up to the SDK to parse the
solidity custom error selectors into understandable an error enum as
new ones are added