feat: add multiple cancel params
This commit adds a helper function for getting the cancel params for
canceling multiple nonces at once. It just groups the nonces by word and
bitwise ORs the masks for each cancelation to get the minimal set of
calls. An integrator should take each CancelParams and use it to call
the invalidateUnorderedNonces function on permit2