fix: generic order parsing
Previously the generic order parser was looking at an absolute offset
into the serialized order. This works only for orders that have the same
number of bytes between their `head` and `tail`. The correct way to do
this lookup is to find the `tail offset` inside the head, which points
to a byte offset that the dynamic sized fields are. Then secondarly
parse the reactor address based on this offset.

for reference around this, check out the generic quoter from uniswapx:
https://github.com/Uniswap/UniswapX/blob/a23f3a3e6ae872b3fcb2857f6ce73f4003f582a0/src/lens/OrderQuoter.sol#L38-L39