feat(sdk-core): support zksync on V3 (#34)
## Description

We have deployed V3 to to ZkSync. We are updating the SDKs to support V3 in FE and BE. First step is to update sdk-core

## How Has This Been Tested?

unit tests

## Are there any breaking changes?

No

## (Optional) Feedback Focus

No

## (Optional) Follow Ups

We will have to create v2-sdk and v3-sdk and universal-router-sdk changes separately, because we still publish individual sdk packages to npm repo.