feature(flashtestations-sdk): accept array of values for mrTd and rtMr0 measurement registers
This is a workaround for the fact that the GCP TEE devices we use do not have a 100%
transparent build and attestation process. In particular we cannot specify the firmware
version of the GCP TEE that will be used when we build the VM image.  This causes problems because GCP TEE's
with different firmware versions will produce different mrTd and rtMr0 values. Since
we can't at build time know what these mrTd and rtMr0 values will be, we need to generate
multiple images and thus multiple mrTd and rtMr0 values, which we pass to the flashtestations-sdk
as arrays. Then, when we want to verify that a given block was built using a workload ID that was derived
from a particular flashbots-images source code, we need to these arrays of measurement registers, find the
cross product of all of them, and see if at least 1 of the resulting workload IDs matches the workload ID
from the block.

Individual commits:

machine-refactor type name from WorkloadMeasureRegisters -> WorkloadMeasurementRegisters
I used the incorrect name previously, now it's the correct type name

split WorkloadMeasurementRegisters type into 2 types

WorkloadMeasurementRegisters can take both a singular and array type value for rtMr0 + mrTd, and SingularWorkloadMeasurementRegisters is for
strictly the singular field type

add more tests